Get all ZIP codes that fall within a given congressional district

Get all ZIP codes that fall within a given congressional district

cd_to_zip(state_fips_code, congressional_district)

Arguments

state_fips_code
A two-digit U.S. FIPS code for a state

congressional_district
A two digit number specifying a congressional district in a given

Value

dataframe of all congressional districts found for given ZIP code, including state code

Examples

cd_to_zip('34','03')

#> 63 ZIP codes found for congressional district 03 

#>      ZIP state_fips congressional_district
#> 1  07731         34                     03
#> 2  08004         34                     03
#> 3  08005         34                     03
#> 4  08010         34                     03
#> 5  08011         34                     03
#> 6  08015         34                     03
#> 7  08016         34                     03
#> 8  08019         34                     03
#> 9  08022         34                     03
#> 10 08036         34                     03
#> 11 08041         34                     03
#> 12 08042         34                     03
#> 13 08046         34                     03
#> 14 08048         34                     03
#> 15 08050         34                     03
#> 16 08052         34                     03
#> 17 08053         34                     03
#> 18 08054         34                     03
#> 19 08055         34                     03
#> 20 08057         34                     03
#> 21 08060         34                     03
#> 22 08064         34                     03
#> 23 08068         34                     03
#> 24 08073         34                     03
#> 25 08075         34                     03
#> 26 08077         34                     03
#> 27 08088         34                     03
#> 28 08091         34                     03
#> 29 08092         34                     03
#> 30 08501         34                     03
#> 31 08505         34                     03
#> 32 08511         34                     03
#> 33 08515         34                     03
#> 34 08518         34                     03
#> 35 08533         34                     03
#> 36 08554         34                     03
#> 37 08562         34                     03
#> 38 08610         34                     03
#> 39 08620         34                     03
#> 40 08640         34                     03
#> 41 08641         34                     03
#> 42 08701         34                     03
#> 43 08721         34                     03
#> 44 08722         34                     03
#> 45 08723         34                     03
#> 46 08724         34                     03
#> 47 08731         34                     03
#> 48 08732         34                     03
#> 49 08734         34                     03
#> 50 08735         34                     03
#> 51 08738         34                     03
#> 52 08739         34                     03
#> 53 08740         34                     03
#> 54 08741         34                     03
#> 55 08742         34                     03
#> 56 08751         34                     03
#> 57 08752         34                     03
#> 58 08753         34                     03
#> 59 08754         34                     03
#> 60 08755         34                     03
#> 61 08757         34                     03
#> 62 08758         34                     03
#> 63 08759         34                     03

cd_to_zip('36','05')

#> 39 ZIP codes found for congressional district 05 

#>      ZIP state_fips congressional_district
#> 1  11001         36                     05
#> 2  11003         36                     05
#> 3  11010         36                     05
#> 4  11022         36                     05
#> 5  11096         36                     05
#> 6  11352         36                     05
#> 7  11366         36                     05
#> 8  11405         36                     05
#> 9  11411         36                     05
#> 10 11412         36                     05
#> 11 11413         36                     05
#> 12 11416         36                     05
#> 13 11417         36                     05
#> 14 11418         36                     05
#> 15 11419         36                     05
#> 16 11420         36                     05
#> 17 11421         36                     05
#> 18 11422         36                     05
#> 19 11423         36                     05
#> 20 11426         36                     05
#> 21 11427         36                     05
#> 22 11428         36                     05
#> 23 11429         36                     05
#> 24 11430         36                     05
#> 25 11432         36                     05
#> 26 11433         36                     05
#> 27 11434         36                     05
#> 28 11435         36                     05
#> 29 11436         36                     05
#> 30 11437         36                     05
#> 31 11451         36                     05
#> 32 11559         36                     05
#> 33 11580         36                     05
#> 34 11581         36                     05
#> 35 11691         36                     05
#> 36 11692         36                     05
#> 37 11693         36                     05
#> 38 11694         36                     05
#> 39 11697         36                     05