Search ZIP codes for a given city within a state

Search ZIP codes for a given city within a state

search_city(city_name, state_abb)

Arguments

city_name
Name of major city to search

state_abb
Two-digit code for a U.S. state

Value

tibble of all ZIP code data found for given city

Examples

search_city('Spring Lake','NJ')

#> 1 ZIP codes found for Spring Lake , NJ 

#>   zipcode zipcode_type  major_city post_office_city
#> 1   07762     Standard Spring Lake  Spring Lake, NJ
#>                                                                             common_city_list
#> 1 78, 9c, 8b, 56, 0a, 2e, 28, ca, cc, 4b, 57, f0, 49, cc, 4e, 55, 8a, 05, 00, 28, d6, 05, 0d
#>            county state   lat    lng timezone radius_in_miles
#> 1 Monmouth County    NJ 40.15 -74.04  Eastern               1
#>                                                                           area_code_list
#> 1 78, 9c, 8b, 56, 32, 37, 36, 52, d2, 51, 50, b2, 30, b1, 50, 8a, 05, 00, 14, ea, 02, cd
#>   population population_density land_area_in_sqmi water_area_in_sqmi
#> 1       8403               2997               2.8               0.42
#>   housing_units occupied_housing_units median_home_value
#> 1          5480                   3975            649400
#>   median_household_income bounds_west bounds_east bounds_north bounds_south
#> 1                   73986   -74.05861   -74.01508     40.16901     40.13751

search_city('Chappaqua','NY')

#> 1 ZIP codes found for Chappaqua , NY 

#>   zipcode zipcode_type major_city post_office_city
#> 1   10514     Standard  Chappaqua    Chappaqua, NY
#>                                                                     common_city_list
#> 1 78, 9c, 8b, 56, 72, ce, 48, 2c, 28, 48, 2c, 2c, 4d, 54, 8a, 05, 00, 1f, 41, 04, 91
#>               county state   lat    lng timezone radius_in_miles
#> 1 Westchester County    NY 41.18 -73.77  Eastern               4
#>                                               area_code_list population
#> 1 78, 9c, 8b, 56, b2, 34, 34, 51, 8a, 05, 00, 06, 6e, 01, 9b      11946
#>   population_density land_area_in_sqmi water_area_in_sqmi housing_units
#> 1                932             12.82               0.13          4033
#>   occupied_housing_units median_home_value median_household_income bounds_west
#> 1                   3893            912700                  202414   -73.81274
#>   bounds_east bounds_north bounds_south
#> 1   -73.72405     41.21839     41.12929